testing

+++ fv3/module_fv3_config.F90
@@ -37,7 +37,7 @@
   !> Output frequency if this array has only two elements and the value of
   !>  the second eletment is -1. Otherwise, it is the specific output forecast
   !>  hours
-  real,dimension(:),allocatable                   :: output_fh
+  real(8),dimension(:),allocatable                   :: output_fh
 
   !> Calendar type
   character(17)            :: calendar='                 '
